pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "disaster_tweets"
version = "0.0.1"
description = "Task 1: Kaggle Disaster Tweets"
authors = [
  { name = "Katsiaryna" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [    #remove locked versions???
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "pip",
    "pycountry>=24.6.1",
    "pyspellchecker>=0.8.4",
    "pytest",
    "python-dotenv",
    "rapidfuzz>=3.14.3",
    "ruff",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "spacy>=3.8.11",
]
requires-python = "~=3.11.0"


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

